refactor(package): move model-related utility classes into model/utils
