+++ app/src/main/java/com/example/ezvault/view/EditItemDetails.java
@@ -47,7 +47,7 @@ import com.example.ezvault.data.database.ImageDAO;
 import com.example.ezvault.data.database.ItemDAO;
 import com.example.ezvault.model.Image;
 import com.example.ezvault.model.Item;
-import com.example.ezvault.model.SerialPrediction;
+import com.example.ezvault.model.utils.SerialPrediction;
 import com.example.ezvault.model.SerialPredictor;
 import com.example.ezvault.utils.textwatchers.NonEmptyTextWatcher;
 import com.example.ezvault.upcAPI;
