+++ app/src/main/java/com/example/ezvault/viewmodel/AddItemFragment.java
@@ -42,9 +42,9 @@ import com.example.ezvault.data.database.ItemDAO;
 import com.example.ezvault.data.database.RawUserDAO;
 import com.example.ezvault.model.Image;
 import com.example.ezvault.model.Item;
-import com.example.ezvault.model.ItemBuilder;
+import com.example.ezvault.model.utils.ItemBuilder;
 import com.example.ezvault.model.ItemList;
-import com.example.ezvault.model.SerialPrediction;
+import com.example.ezvault.model.utils.SerialPrediction;
 import com.example.ezvault.model.SerialPredictor;
 import com.example.ezvault.model.Tag;
 import com.example.ezvault.model.User;
