+++ app/src/test/java/com/example/ezvault/ItemBuilderTest.java
@@ -5,7 +5,7 @@ import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertTrue;
 
 import com.example.ezvault.model.Item;
-import com.example.ezvault.model.ItemBuilder;
+import com.example.ezvault.model.utils.ItemBuilder;
 
 import org.junit.Test;
